remove strings file
