+++ src/platform/plugins/private/presentation_panel/public/panel_component/use_panel_error_css.ts
@@ -17,6 +17,6 @@ export const usePanelErrorCss = () => {
     '& > *': {
       maxHeight: '100%',
       overflow: 'auto',
-    }
+    },
   });
-}
+};
